+++ config/connection.js
@@ -1,0 +1,38 @@
+const Sequelize = require("sequelize");
+const session = require("express-session");
+const SequelizeStore = require("connect-session-sequelize")(session.Store);
+require("dotenv").config();
+
+let sequelize;
+
+if (process.env.JAWSDB_URL) {
+  sequelize = new Sequelize(process.env.JAWSDB_URL);
+} else {
+  sequelize = new Sequelize(
+    process.env.DB_NAME,
+    process.env.DB_USER,
+    process.env.DB_PASSWORD,
+    {
+      host: "localhost",
+      dialect: "mysql",
+      port: 3305,
+    }
+  );
+}
+
+const sessionConfig = {
+  secret: process.env.SECRET,
+  cookie: {
+    maxAge: 1 * 60 * 60 * 1000, // 1 * 60 * 60 * 1000 === 1 hour
+    httpOnly: true,
+    secure: false,
+    sameSite: "strict",
+  },
+  resave: false,
+  saveUninitialized: true,
+  store: new SequelizeStore({
+    db: sequelize,
+  }),
+};
+
+module.exports = { sequelize, sessionConfig };
