+++ models/Login.js
@@ -1,0 +1,45 @@
+const { Model, DataTypes } = require("sequelize");
+const bcrypt = require("bcrypt");
+const { sequelize } = require("../config/connection");
+
+class Login extends Model {
+  checkPassword(loginPw) {
+    return bcrypt.compareSync(loginPw, this.password);
+  }
+}
+
+User.init(
+  {
+    username: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      unique:true,
+    },
+    email: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      unique: true,
+      validate: {
+        isEmail: true,
+      },
+    },
+    password: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      validate: {
+        len: [8],
+      },
+    },
+  },
+  {
+    hooks: {
+      beforeCreate: async (newUserData) => {
+        newUserData.password = await bcrypt.hash(newUserData.password, 10);
+        return newUserData;
+      },
+    },
+    sequelize,
+  }
+);
+
+module.exports = Login;
